+++ source/solvers/heat_transfer.cc
@@ -1574,7 +1574,7 @@ HeatTransfer<dim>::solve_linear_system(const bool initial_step,
                   << linear_solver_tolerance << std::endl;
     }
 
-  const double ilu_fill =
+  const unsigned int ilu_fill =
     simulation_parameters.linear_solver.at(PhysicsID::heat_transfer)
       .ilu_precond_fill;
   const double ilu_atol =
